refactor: widget으로 분리
